+++ server/lib/src/router/api/api.dart
@@ -7,7 +7,9 @@ import '../router_interface.dart';
 import 'auth/auth.dart';
 import 'challenge/challenge.dart';
 import 'game/game.dart';
+import 'games/games.dart';
 import 'player/player.dart';
+import 'players/players.dart';
 import 'settings/settings.dart';
 
 /// Router that handles requests on `/ICServerRouter.api`.
@@ -15,8 +17,11 @@ class ApiRouter implements RouterInterface {
   /// Router that handles auth requests.
   final AuthRouter _authRouter;
 
-  /// Router that handles requests regarding players.
+  /// Router that handles requests for single player.
   final PlayerRouter _playerRouter;
+
+  /// Router that handles requests regarding players.
+  final PlayersRouter _playersRouter;
 
   /// Router that handles requests for games.
   final ChallengeRouter _challengeRouter;
@@ -24,8 +29,11 @@ class ApiRouter implements RouterInterface {
   /// Router that handles requests for settings.
   final SettingsRouter _settingsRouter;
 
-  /// Router that handles requests for games.
+  /// Router that handles requests for a single game.
   final GameRouter _gameRouter;
+
+  /// Router that handles requests for games.
+  final GamesRouter _gamesRouter;
 
   /// Constructs new `ApiRouter` with given [authRouter], [playerRouter],
   /// [challengeRouter], [settingsRouter], and [gameRouter].
@@ -34,14 +42,18 @@ class ApiRouter implements RouterInterface {
   ApiRouter({
     AuthRouter? authRouter,
     PlayerRouter? playerRouter,
+    PlayersRouter? playersRouter,
     ChallengeRouter? challengeRouter,
     SettingsRouter? settingsRouter,
     GameRouter? gameRouter,
+    GamesRouter? gamesRouter,
   })  : _authRouter = authRouter ?? AuthRouter(),
         _playerRouter = playerRouter ?? PlayerRouter(),
+        _playersRouter = playersRouter ?? PlayersRouter(),
         _challengeRouter = challengeRouter ?? ChallengeRouter(),
         _settingsRouter = settingsRouter ?? SettingsRouter(),
-        _gameRouter = gameRouter ?? const GameRouter();
+        _gameRouter = gameRouter ?? const GameRouter(),
+        _gamesRouter = gamesRouter ?? const GamesRouter();
 
   /// Request handler / rerouter.
   @override
@@ -57,12 +69,16 @@ class ApiRouter implements RouterInterface {
         return await _authRouter.handle(request);
       case ICServerRoute.apiPlayer:
         return await _playerRouter.handle(request);
+      case ICServerRoute.apiPlayers:
+        return await _playersRouter.handle(request);
       case ICServerRoute.apiChallenge:
         return await _challengeRouter.handle(request);
       case ICServerRoute.apiSettings:
         return await _settingsRouter.handle(request);
       case ICServerRoute.apiGame:
         return await _gameRouter.handle(request);
+      case ICServerRoute.apiGames:
+        return await _gamesRouter.handle(request);
       default:
         return respondWithBadRequest(request);
     }